Rename Childcare V2 to just childcare
Change status to published.
