--- lib/flows/locales/en/childcare-costs-for-tax-credits-v2.yml
@@ -1,191 +1,0 @@
-en-GB:
-  flow:
-    childcare-costs-for-tax-credits-v2:
-      phrases:
-        cost_change_doesnt_matter: "This doesn’t affect your tax credits. You don’t need to report it."
-        cost_change_does_matter: "Tell the Tax Credit Office so they can adjust your tax credit."
-        cost_change_4_weeks: |
-          Call the Tax Credit Office if you expect this to last for 4 weeks in a row or more.
-
-          You don’t need to report it if this won’t last 4 weeks in a row or more.
-
-      options:
-        "yes": "Yes"
-        "no": "No"
-        weekly_same_amount: "Weekly, and I always pay the same amount"
-        weekly_diff_amount: "Weekly, and the amount I pay varies"
-        monthly_same_amount: "Monthly, and I always pay the same amount"
-        monthly_diff_amount: "Monthly, and the amount I pay varies"
-        other: "Other"
-
-      title: "Tax credits: working out your childcare costs"
-      body: |
-        You can claim extra tax credits to [help with your childcare costs](/childcare-tax-credits "Childcare and tax credits") if you’re eligible. Use this tool to work out what childcare costs you should claim. You can make your claim 7 days before you start paying for childcare.
-
-        If you already claim tax credits for childcare costs and your costs have changed, you can use this tool to work out if you need to report the change to the Tax Credit Office.
-
-        ^When calculating your childcare costs, you can only include costs you pay yourself. You may be fined if you provide the wrong amount.^
-
-        You cannot include childcare:
-
-          - payments from your employer - either in money or childcare vouchers  
-
-          - vouchers in return for a reduction in your pay - this is called a 'salary sacrifice'  
-
-          - payments or grants from a government scheme, for example to help you start work or because you are a student  
-
-          - costs your educational or local authority is paying, for example for your child's nursery childcare  
-
-      #Q1
-      currently_claiming?:
-        title: "Are you currently claiming tax credits for childcare costs?"
-
-      #Q2
-      how_often_use_childcare?:
-        title: "How often do you use childcare?"
-        options:
-          regularly_less_than_year: "I’ve been using childcare regularly for less than a year"
-          regularly_more_than_year: "I've been using childcare regularly for a year or more"
-          only_short_while: "I only use childcare for short periods once in a while"
-
-      #Q3
-      have_costs_changed?:
-        title: "Have the costs of your childcare changed since you last made your claim?"
-
-      #Q4
-      how_often_pay_1?:
-        title: "How often do you pay your childcare provider(s)?"
-
-      #Q5
-      how_often_pay_2?:
-        title: "How often do you pay your childcare provider(s)?"
-
-      #Q6
-      how_much_12_months_1?:
-        title: "How much do you expect to pay in total for childcare for the next 12 months?"
-        hint: "Add up the amounts you expect to pay in total for the next 12 months - start from the day you’re doing the calculation"
-
-      #Q7
-      how_much_52_weeks_1?:
-        title: "How much do you expect to pay in total for childcare for the next 52 weeks?"
-        hint: "Add up the amounts you expect to pay in total for the next 52 weeks - start from the day you’re doing the calculation"
-
-      #Q8
-      how_much_52_weeks_2?:
-        title: "How much do you expect to pay in total for childcare for the next 52 weeks?"
-
-      #Q9
-      how_much_12_months_2?:
-        title: "How much do you expect to pay in total for childcare for the next 12 months?"
-
-      #Q10
-      how_much_each_month?:
-        title: "How much do you pay each month?"
-
-      #Q11
-      pay_same_each_time?:
-        title: "Do you pay the same each time?"
-        hint: |
-          Sometimes you may pay - or expect to pay - different amounts for childcare (known as ‘variable costs’), for example:
-
-          - you regularly use childcare, but pay more during school holidays than you do at term time
-          - the hours you use childcare change from week to week or month to month
-
-      #Q12
-      how_often_pay_providers?:
-        title: "How often do you pay your childcare provider(s)?"
-        options:
-          weekly: "Weekly"
-          fortnightly: "Fortnightly"
-          every_4_weeks: "Every 4 weeks"
-          every_month: "Every calendar month"
-          termly: "Termly"
-          yearly: "Yearly"
-
-      #Q13
-      how_much_fortnightly?:
-        title: "How much do you pay each fortnight?"
-
-      #Q14
-      how_much_4_weeks?:
-        title: "How much do you pay every 4 weeks?"
-
-      #Q15
-      how_much_yearly?:
-        title: "How much do you pay every year?"
-
-      #Q16
-      how_much_spent_last_12_months?:
-        title: "How much have you spent on childcare in the last 12 months?"
-        hint: "Add up how much you have spent in total over the last 12 months - start backwards from the date you’re doing the calculation."
-
-      #Q17
-      new_weekly_costs?:
-        title: "What are your new weekly costs?"
-        body: "Round the total up to the nearest pound."
-
-      #Q18
-      old_weekly_amount_1?:
-        title: "What was the old average weekly amount you gave the Tax Credit Office?"
-        body: |
-          If you don’t know what your previous childcare costs were, check your award or renewal notice.
-
-          Round the total up to the nearest pound.
-      #Q19
-      new_monthly_cost?:
-        title: "What's the new monthly cost of your childcare?"
-
-      #Q20
-      old_weekly_amount_2?:
-        title: "What was the old average weekly amount you gave the Tax Credit Office?"
-        body: |
-          If you don’t know what your previous childcare costs were, check your award or renewal notice.
-
-          Round the total up to the nearest pound.
-
-      #Q21
-      old_weekly_amount_3?:
-        title: "What was the old average weekly amount you gave the Tax Credit Office?"
-        body: |
-          If you don’t know what your previous childcare costs were, check your award or renewal notice.
-
-          Round the total up to the nearest pound.
-
-      # OUTCOMES
-      call_helpline_detailed:
-        title: "Call the Tax Credit Helpline to work out your average weekly childcare costs."
-        body: |
-          You’ll need to tell them:
-
-          - the start and end date of the childcare
-
-          - how much you spent on childcare for the short period
-          - your childcare provider’s details - their name, address, phone number, childcare registration or approval number
-
-      call_helpline_plain:
-        title: "Contact the Tax Credit Office for help in working out your costs."
-
-      no_change:
-        title: "There is no change to your tax credits."
-        body: "You don’t have to report anything to the Tax Credit Office."
-
-      round_up_weekly:
-        title: "Your weekly childcare costs are the total amount you pay each week rounded up to the nearest pound."
-        body: "Use this figure on your claim form."
-      weekly_costs_are_x:
-        title: "Your weekly childcare costs are £%{weekly_cost}."
-        body: "Use this figure on your claim form."
-
-      cost_changed:
-        title: "Your average weekly childcare costs have %{title_change_text} by %{difference_money}."
-        body: |
-          %{body_phrases}
-
-      no_longer_paying:
-        title: "You’re no longer paying for childcare."
-        body: |
-          If you expect this to last for 4 weeks or more, tell the [Tax Credit Office](/contact-the-tax-credit-office "Contact the Tax Credit Office").
-
-
-
-
