+++ test/integration/flows/childcare_costs_for_tax_credits_test.rb
@@ -2,404 +2,322 @@
 require_relative '../../test_helper'
 require_relative 'flow_test_helper'
 
-class ChildcareCostsForTaxCreditsTest < ActiveSupport::TestCase
+class ChildcareCostsForTaxCreditsV2Test < ActiveSupport::TestCase
   include FlowTestHelper
-  
+
   setup do
     setup_for_testing_flow 'childcare-costs-for-tax-credits'
   end
-  
-  # Q1
-  context "ask 'is the the first time you have claimed?'" do
-    should "be true" do
-      assert_current_node :first_time_claim?
-    end
-    
-    context "answer 'yes'" do
+
+  context "answering Q1" do
+    context "answering with yes" do
       setup do
         add_response :yes
       end
-      # Q2
-      should "ask 'about your childcare...'" do
-        assert_current_node :frequency_of_childcare?
+
+      should "take you to Q3 if you answer yes" do
+        assert_current_node :have_costs_changed?
       end
-      
-      context "answer 'regularly for less than a year'" do
-        setup do
-          add_response :regularly_less_than_a_year
+
+      context "answering Q3" do
+        should "take you to outcome if you answer no" do
+          add_response :no
+          assert_current_node :no_change
         end
-        # Q4
-        should "ask 'how often do you pay your providers?'" do
-          assert_current_node :how_often_do_you_pay?
+
+        should "take you to how_often_pay_2 if you answer yes" do
+          add_response :yes
+          assert_current_node :how_often_pay_2?
         end
-        context "answer 'same amount weekly'" do
-          # A1
-          should "say 'round up the total'" do
-            add_response :same_amount_weekly
-            assert_current_node :round_up_total
-          end
-        end
-        context "answer 'varying amount weekly'" do
-          setup do
-            add_response :varying_amount_weekly
-          end
-          # C1
-          should "ask 'what is the total for 52 weeks?'" do
-            assert_current_node :costs_for_year_in_weeks?
-          end
-          context "answer '4000'" do
-            setup do
-              add_response 4000
-            end
-            # A2
-            should "say 'Your weekly childcare costs are 77'" do
-              assert_current_node :weekly_costs
-              assert_state_variable "cost", 77
-            end
-          end
-        end
-        context "answer 'same amount monthly" do
-          setup do
-            add_response :same_monthly
-          end
-          # C3
-          should "ask 'how much do you pay each month?'" do
-            assert_current_node :how_much_do_you_pay_each_month?
-          end
-          context "answer 350" do
-            setup do
-              add_response 350
-            end
-            # A4
-            should "say 'Your  weekly childcare costs are 81'" do
-              assert_current_node :weekly_costs
-              assert_state_variable "cost", 81
-            end
-          end
-        end
-        context "answer 'varying amount monthly'" do
-          setup do
-            add_response :varying_amount_monthly
-          end
-          # C2
-          should "ask 'what is the total cost over 12 months?'" do
-            assert_current_node :costs_for_year_in_months?
-          end
-          context "answer '4200'" do
-            setup do
-              add_response 4200
-            end
-            # A3
-            should "say 'Your weekly childcare costs are 81'" do
-              assert_current_node :weekly_costs
-              assert_state_variable "cost", 81
-            end
-          end
-        end
-        context "answer 'other'" do
-          setup do
-            add_response :other
-          end
-          # C4
-          should "ask 'what is the total cost for 12 months?'" do
-            assert_current_node :costs_for_year_in_months?
-          end
-          context "answer 3800" do
-            setup do
-              add_response 3800
-            end
-            # A5
-            should "say 'Your weekly childcare cost is 74'" do
-              assert_current_node :weekly_costs
-              assert_state_variable "cost", 74
-            end
-          end
-        end
-      end
-      
-      context "answer 'regularly for more than a year'" do
-        setup do
-          add_response :regularly_more_than_a_year
-        end
-        # Q5
-        should "ask 'do you pay the same every time?'" do
-          assert_current_node :do_you_pay_the_same_every_time?
-        end
-        context "answer 'yes'" do
+
+        context "answering Q5" do
           setup do
             add_response :yes
           end
-          # Q6
-          should "ask 'how often do you pay your providers?'" do
-            assert_current_node :how_often_do_you_pay_your_providers?
+
+          should "be Q5" do
+            assert_current_node :how_often_pay_2?
           end
-          context "answer 'weekly'" do
-            should "say 'round up the total cost'" do
-              add_response :weekly
-              assert_current_node :round_up_total
-            end
+
+          should "take you to weekly costs for weekly answer" do
+            add_response :weekly_same_amount
+            assert_current_node :new_weekly_costs?
           end
-          context "answer 'fortnightly'" do
-            setup do
-              add_response :fornightly
-            end
-            # C5
-            should "ask 'how much do you pay each fortnight?'" do
-              assert_current_node :how_much_do_you_pay_each_fortnight?
-            end
-            context "answer 90" do
-              setup do
-                add_response 90
-              end
-              # A7
-              should "say 'Your weekly costs are 45 use this amount on your claim form'" do
-                assert_current_node :weekly_costs_for_claim_form
-                assert_state_variable "cost", 45
-              end
-            end
+
+          should "take you to how_much_12_months_2 if you answer with weekly diff amount" do
+            add_response :weekly_diff_amount
+            assert_current_node :how_much_52_weeks_2?
           end
-          context "answer 'every 4 weeks'" do
-            setup do
-              add_response :every_four_weeks
-            end
-            # C6
-            should "ask 'how much do you pay every 4 weeks?'" do
-              assert_current_node :how_much_do_you_pay_every_four_weeks?
-            end
-            context "answer 128" do
-              setup do
-                add_response 128
-              end
-              # A8
-              should "say 'Your weekly costs are 32 use this amount on your claim form'" do
-                assert_current_node :weekly_costs_for_claim_form
-                assert_state_variable "cost", 32
-              end
-            end
+
+          should "take you to the new_monthly_cost if you say monthly_same_amount" do
+            add_response :monthly_same_amount
+            assert_current_node :new_monthly_cost?
           end
-          context "answer 'monthly'" do
-            setup do
-              add_response :monthly
-            end
-            # C7
-            should "ask 'how much do you pay each month?'" do
-              assert_current_node :how_much_do_you_pay_each_month?
-            end
+
+          should "take you to 52 weeks question if you answer with monthly diff" do
+            add_response :monthly_diff_amount
+            assert_current_node :how_much_12_months_2?
           end
-          context "answer 'termly'" do
-            # A10
-            should "say 'contact the tax credit office'" do
-              add_response :termly
-              assert_current_node :contact_the_tax_credit_office
-            end
-          end
-          context "answer 'yearly'" do
-            setup do
-              add_response :yearly
-            end
-            # C8
-            should "ask 'how much do you pay annually?'" do
-              assert_current_node :how_much_do_you_pay_anually?
-            end
-            context "answer '4200'" do
-              setup do
-                add_response 4200
-              end
-              
-              should "say 'Your weeks costs are 81'" do
-                assert_current_node :weekly_costs_for_claim_form
-                assert_state_variable "cost", 81
-              end
-            end
-          end
-          context "answer 'other'" do
-            # A12
-            should "say 'contact the tax credit office'" do
-              add_response :other
-              assert_current_node :contact_the_tax_credit_office
-            end
-          end
-        end
-        context "answer 'no'" do
-          setup do
-            add_response :no
-          end
-          # C9
-          should "ask 'what is your annual cost?'" do
-            assert_current_node :varying_annual_cost?
-          end
-          context "answer '3950'" do
-            setup do
-              add_response 3950
-            end
-            
-            should "say 'Your weeks costs are 76'" do
-              assert_current_node :weekly_costs_for_claim_form
-              assert_state_variable "cost", 76
-            end
-          end
-        end
-      end
-      context "answer 'intermittently'" do
-        # A19
-        should "say 'call the helpline'" do
-          add_response :intermittently
-          assert_current_node :call_the_helpline
-        end
-      end
-      
+        end #Q5
+
+
+      end #Q3
     end
-    
-    context "answer 'no'" do
+
+    context "answering with no" do
       setup do
         add_response :no
       end
-      # Q3
-      should "ask 'have the costs changed?'" do
-        assert_current_node :have_the_costs_changed?
+
+      should "take you to Q2 if you answer no" do
+        assert_current_node :how_often_use_childcare?
       end
-      context "answer 'yes'" do
-        setup do
-          add_response :yes
+
+      context "answering Q2" do
+        should "take you to how_often_pay_1 if you answer less than year" do
+          add_response :regularly_less_than_year
+          assert_current_node :how_often_pay_1?
         end
-        # Q7
-        should "ask 'how often and what do you pay providers?'" do
-          assert_current_node :how_often_and_what_do_you_pay_your_providers?
+
+        should "take you to pay_same_each_time if you answer more than year" do
+          add_response :regularly_more_than_year
+          assert_current_node :pay_same_each_time?
         end
-        context "answer 'same amount weekly'" do
+
+        should "take you to outcome if you answer only_short_while" do
+          add_response :only_short_while
+          assert_current_node :call_helpline_detailed
+        end
+
+        context "answering Q4" do
           setup do
-            add_response :same_amount_weekly # => :old_weekly_costs? # C10
+            add_response :regularly_less_than_year
           end
-          # C10A
-          should "ask 'what do you expect your new weekly costs to be?'" do
-            assert_current_node :new_weekly_costs?
+
+          should "take you to round_up_weekly outcome if you answer weekly_same_amount" do
+            add_response :weekly_same_amount
+            assert_current_node :round_up_weekly
           end
-          context "answer '60'" do
+
+          should "take you to how_much_12_months if you answer with other" do
+            add_response :other
+            assert_current_node :how_much_12_months_1?
+          end
+
+          context "answering Q7" do
             setup do
-              add_response 60
+              add_response :weekly_diff_amount
             end
-            # C10B
-            should "ask 'what were your previous costs?'" do
-              assert_current_node :old_weekly_costs?
+            should "calculate the weekly cost and take user to outcome" do
+              add_response 52
+              assert_current_node :weekly_costs_are_x
+              assert_state_variable :weekly_cost, 1
             end
-            # A14 where diff is >= 10
-            context "answer 45" do
-              setup do
-                add_response 45
-              end              
-              should "say 'costs have increased'" do
-                assert_current_node :costs_have_increased
-              end
-            end
-            # A14 where diff < 10
-            context "answer 55" do
-              setup do
-                add_response 55
-              end
-              should "say 'costs have not increased'" do
-                assert_current_node :costs_have_increased_below_threshold
-              end
-            end
-          end
-        end
-        context "answer 'varying amount weekly'" do
-          setup do
-            add_response :varying_amount_weekly # => :old_annual_costs? # C11
-          end
-          # C11A
-          should "ask 'what is your expected annual cost?'" do
-            assert_current_node :new_annual_costs?
-          end
-          context "answer 3500" do
+
+          end #Q7
+
+          context "answering Q6" do
             setup do
-              add_response 3500
+              add_response :other # answer Q4
+              add_response 52 # answer Q6
             end
-            # C11B
-            should "ask 'what were your previous weekly costs?'" do
-              assert_current_node :old_annual_costs?
+            should "take you to weekly costs outcome" do
+              assert_current_node :weekly_costs_are_x
             end
-            context "answer '58' (10 below current)" do
-              setup do
-                add_response 58
-              end
-              # A15
-              should "say 'your costs have increased'" do
-                assert_current_node :costs_have_increased
-              end
+
+            should "calculate the weekly cost" do
+              assert_state_variable :weekly_cost, 1
             end
-            context "answer '59' (9 below current)" do
-              setup do
-                add_response 59
-              end
-              # A15
-              should "say 'your costs have not increased enough'" do
-                assert_current_node :costs_have_increased_below_threshold
-              end
-            end
-            context "answer '69'" do
-              setup do
-                add_response 69
-              end
-              # A15
-              should "say 'your costs have not decreased enough'" do
-                assert_current_node :costs_have_decreased_below_threshold
-              end
-            end
-            context "answer '68'" do
-              setup do
-                add_response 68
-              end
-              # A21
-              should "say 'your costs have not increased enough'" do
-                assert_current_node :no_change_to_credits
-              end
-            end
-            context "answer '90'" do
-              setup do
-                add_response 90
-              end
-              # A15
-              should "say 'your costs have decreased'" do
-                assert_current_node :costs_have_decreased
-              end
-            end
-          end
-        end
-        context "answer 'same amount monthly'" do
-          setup do
-            add_response :same_monthly # => :old_average_weekly_costs? # C13
-          end
-          # C13
-          should "ask 'what were your previous average weekly costs?'" do
-            assert_current_node :new_average_weekly_costs?
-          end
-        end
-        context "answer 'varying amount monthly'" do
-          setup do
-            add_response :varying_amount_monthly # => :old_annual_costs? # C14
-          end
-          # C14
-          should "ask 'what were your previous annual costs?'" do
-            assert_current_node :new_annual_costs? # C12
-          end
-        end
-        context "answer 'other'" do
-          setup do
-            add_response :other # => :old_annual_costs? # C12
-          end
-          # C12
-          should "ask 'what were your previous annual costs?'" do
-            assert_current_node :new_annual_costs?
-          end
-        end
+
+          end #Q6
+
+        end #Q4
+
+      end #Q2
+
+    end
+  end #Q1
+
+  context "calculating weekly costs" do
+    context "through Question 10" do
+      setup do
+        add_response :no #Q1
+        add_response :regularly_more_than_year #Q2
+        add_response :yes #Q11
+        add_response :every_month #Q12
       end
-      context "answer 'no'" do
-        # A20
-        should "say 'no change to tax credits" do
-          add_response :no
-          assert_current_node :no_change_to_credits
-        end
+
+      should "calculate the weekly cost" do
+        add_response 4 #Q10
+        assert_state_variable :weekly_cost, 1
+        assert_current_node :weekly_costs_are_x
       end
     end
+    context "through Question 13" do
+      setup do
+        add_response :no #Q1
+        add_response :regularly_more_than_year #Q2
+        add_response :yes #Q11
+        add_response :fortnightly #Q12
+      end
+
+      should "ask you how much you pay fortnightly" do
+        add_response 10 # Answer Q13
+        assert_state_variable :weekly_cost, 5
+        assert_current_node :weekly_costs_are_x
+      end
+    end # Q13
+
+    context "through Question 14" do
+      setup do
+        add_response :no #Q1
+        add_response :regularly_more_than_year #Q2
+        add_response :yes #Q11
+        add_response :every_4_weeks #Q12
+      end
+
+      should "calculate the weekly cost" do
+        add_response 20 #Q14
+        assert_state_variable :weekly_cost, 5
+        assert_current_node :weekly_costs_are_x
+      end
+    end # Q14
+
+    context "through Question 15" do
+      setup do
+        add_response :no #Q1
+        add_response :regularly_more_than_year #Q2
+        add_response :yes #Q11
+        add_response :yearly #Q12
+      end
+
+      should "calculate the weekly cost" do
+        add_response 52 #Q14
+        assert_state_variable :weekly_cost, 1
+        assert_current_node :weekly_costs_are_x
+      end
+    end # Q15
+  end # questions that lead to weekly-costs outcome
+
+  context "questions that calculate the difference in costs" do
+    context "Through Question 18" do
+      setup do
+        add_response :yes #Q1
+        add_response :yes #Q3
+        add_response :weekly_diff_amount #Q5
+      end
+
+      should "be at Q8" do
+        assert_current_node :how_much_52_weeks_2?
+      end
+
+      should "be at Q18" do
+        add_response 52 #Q8
+        assert_current_node :old_weekly_amount_1?
+      end
+
+      should "calculate weekly_cost from Q8" do
+        add_response 52 #Q8
+        assert_state_variable :weekly_cost, 1
+      end
+
+      should "calculate diff after answering Q18" do
+        add_response 52 # Q8
+        add_response 2 # Q18
+        assert_state_variable :old_weekly_cost, 2
+        assert_state_variable :weekly_difference, -1
+        assert_state_variable :weekly_difference_abs, 1
+        assert_current_node :cost_changed
+      end
+    end # Q18
+
+  end # questions that calculate cost difference
+
+  context "going through Question 17" do
+    setup do
+      add_response :yes #Q1
+      add_response :yes #Q3
+      add_response :weekly_same_amount #Q5
+    end
+
+    should "be at Q17" do
+      assert_current_node :new_weekly_costs?
+    end
+
+    should "take user to not paying output if they answer 0" do
+      add_response 0
+      assert_current_node :no_longer_paying
+    end
+
+    should "take user to Q20 if they give an answer" do
+      add_response 1
+      assert_state_variable :new_weekly_costs, 1
+      assert_current_node :old_weekly_amount_2?
+    end
+
+    context "answering Q20" do
+      setup do
+        add_response 1 # Q17
+        add_response 11 # Q20
+      end
+
+      should "calculate the old costs based on user answer" do
+        assert_state_variable :old_weekly_costs, 11
+        assert_state_variable :weekly_difference, -10
+        assert_state_variable :weekly_difference_abs, 10
+        assert_state_variable :ten_or_more, true
+        assert_current_node :cost_changed
+      end
+
+      should "show correct phrases" do
+        assert_phrase_list :body_phrases, [:cost_change_4_weeks]
+        assert_state_variable :title_change_text, "decreased"
+      end
+
+    end # Q20
+  end # Q17
+
+  context "going through Q19" do
+    setup do
+      add_response :yes #Q1
+      add_response :yes #Q3
+      add_response :monthly_same_amount #Q5
+    end
+
+    should "be at Q19" do
+      assert_current_node :new_monthly_cost?
+    end
+
+    should "take user to not paying output if they answer 0" do
+      add_response 0
+      assert_current_node :no_longer_paying
+    end
+
+    should "take the user to Q21 if they give an answer" do
+      add_response 4
+      assert_state_variable :new_weekly_costs, 1
+      assert_current_node :old_weekly_amount_3?
+    end
+
+    context "answering Q21" do
+      setup do
+        add_response 4 # Q19
+        add_response 10
+      end
+
+      should "calculate old costs and difference" do
+        assert_state_variable :old_weekly_costs, 10
+        assert_state_variable :weekly_difference, -9
+        assert_state_variable :ten_or_more, false
+        assert_state_variable :title_change_text, "decreased"
+        assert_current_node :cost_changed
+      end
+
+      should "show the correct phrase list" do
+        assert_phrase_list :body_phrases, [:cost_change_doesnt_matter]
+      end
+
+    end
+
   end
 end
